# Reduce ParallelRunnerController CPU usage