--- a/Common/Util/ParallelRunner.cs
+++ b/Common/Util/ParallelRunner.cs
@@ -131,6 +131,10 @@ namespace QuantConnect.Util
                     }
                     else
                     {
+                        if (!_holdQueue.Any(x => x.IsReady))
+                        {
+                            Thread.Sleep(5);
+                        }
                         _holdQueue.Add(workItem, token);
                     }
                 }