@jaredbroad 2019-03-01
Rethinking for bigger picture.